+++ src/static/src/webpack.config.js
@@ -1,5 +1,5 @@
-const webpack = require('webpack');
-const UglifyPlugin = require('uglifyjs-webpack-plugin');
+const TerserPlugin = require('terser-webpack-plugin');
+// const UglifyPlugin = require('uglifyjs-webpack-plugin');
 
 const PROD = JSON.parse(process.env.PROD_ENV || '0');
 
@@ -12,20 +12,37 @@ module.exports = {
         filename: 'framework.min.js'
     },
     mode: 'production',
+    performance: {
+        maxAssetSize: 9999999,
+        assetFilter: (asset) => {
+            return asset.match('framwork.min.js')
+        }
+    },
     optimization: {
+        minimize: true,
         minimizer: [
-            new UglifyPlugin({
-                extractComments: false,
-                uglifyOptions: {
-                    output: {
-                        comments: false,
-                        beautify: false,
-                        
-                    },
-                    mangle: false,
-                    compress: false,
+            new TerserPlugin({
+                extractComments: {
+                    condition: false,
+                },
+                terserOptions: {
+                    compress: true
                 }
             })
         ]
+        //     minimizer: [
+        //         new UglifyPlugin({
+        //             extractComments: false,
+        //             uglifyOptions: {
+        //                 output: {
+        //                     comments: false,
+        //                     beautify: false,
+
+        //                 },
+        //                 mangle: false,
+        //                 compress: false,
+        //             }
+        //         })
+        //     ]
     }
 };